AUDIT-47: Refactor the PasswordAuditAdvice to aop under api
