+++ api/src/main/java/org/openmrs/module/auditlogweb/AuditlogwebActivator.java
@@ -15,7 +15,7 @@ import org.apache.commons.logging.LogFactory;
 import org.openmrs.api.UserService;
 import org.openmrs.api.context.Context;
 import org.openmrs.module.BaseModuleActivator;
-import org.openmrs.module.auditlogweb.advice.PasswordAuditAdvice;
+import org.openmrs.module.auditlogweb.api.aop.PasswordAuditAdvice;
 
 /**
  * This class contains the logic that is run every time this module is either started or shutdown
